build: make the ci work again
While this project didn't receive many updates, the Go and Protobuf
community were evolving and they decided to move the development from
the being under to Go's umbrella to be on the Protobuf's. They also
decided to make the original project in github.com/golang/protobuf just
a wrapper around the new project located in
github.com/protocolbuffers/protobuf-go. This made it possible to include
a lot of fixed but at the same time a lot of breaking changes were
included.

Pinning to a version before doing this major bump on the Golang Protobuf
API (they coined as APIv2) solved the issue.

More info.: https://blog.golang.org/protobuf-apiv2